Changed REST-Method for deleteReservation to DELETE; When a reservation gets deleted, its bill gets cancelled

+++ src/main/java/sese/services/ReservationService.java
@@ -33,18 +33,20 @@ public class ReservationService {
     private CustomerService customerService;
     private ModelMapper modelMapper;
     private MailService mailService;
+    private BillService billService;
 
     @Autowired
     private LogService logService;
 
     @Autowired
-    public ReservationService(CustomerRepository customerRepository, CustomerService customerService, ModelMapper modelMapper, RoomRepository roomRepository, ReservationRepository reservationRepository, MailService mailService) {
+    public ReservationService(CustomerRepository customerRepository, CustomerService customerService, ModelMapper modelMapper, RoomRepository roomRepository, ReservationRepository reservationRepository, MailService mailService, BillService billService) {
         this.customerRepository = customerRepository;
         this.customerService = customerService;
         this.modelMapper = modelMapper;
         this.roomRepository = roomRepository;
         this.reservationRepository = reservationRepository;
         this.mailService = mailService;
+        this.billService = billService;
     }
 
     @Transactional
@@ -201,6 +203,14 @@ public class ReservationService {
     }
 
     public void deleteReservation(Long reservationId) {
-        reservationRepository.deleteById(reservationId);
+        Optional<Reservation> ro = reservationRepository.findById(reservationId);
+
+        if(ro.isPresent()) {
+            Reservation reservation = ro.get();
+            billService.cancleBill(reservation.getBill().getId());
+            reservationRepository.deleteById(reservationId);
+        } else {
+            throw new SeseException(SeseError.RESERVATION_NOT_FOUND);
+        }
     }
 }
